Enhancement: IO functions throw exception instead of returning error code
The current IO functions like the system IO functions in C, they return
an error code to indicate that the IO function has error.
In our storage, the IO operation seems unlikely be able to continue,
the IO error should throw an exception to cancel the current query.

+++ contrib/pax_storage/src/cpp/storage/local_file_system.h
@@ -1,9 +1,10 @@
 #pragma once
 
+#include <sys/stat.h>
+
 #include <string>
 #include <utility>
 #include <vector>
-#include <sys/stat.h>
 
 #include "comm/cbdb_wrappers.h"
 #include "comm/singleton.h"
@@ -20,8 +21,8 @@ class LocalFile final : public File {
 
   ssize_t Read(void *ptr, size_t n) override;
   ssize_t Write(const void *ptr, size_t n) override;
-  ssize_t PWrite(const void *ptr, size_t n, size_t offset) override;
-  ssize_t PRead(void *ptr, size_t n, size_t offset) override;
+  ssize_t PWrite(const void *ptr, size_t n, off_t offset) override;
+  ssize_t PRead(void *ptr, size_t n, off_t offset) override;
   size_t FileLength() const override;
   void Close() override;
   void Flush() override;
@@ -40,12 +41,17 @@ class LocalFileSystem final : public FileSystem {
   File *Open(const std::string &file_path) override;
   std::string BuildPath(const File *file) const override;
   void Delete(const std::string &file_path) const override;
-  std::vector<std::string> ListDirectory(const std::string &path) const override;
-  void CopyFile(const std::string &src_file_path, const std::string &dst_file_path) const override;
+  std::vector<std::string> ListDirectory(
+      const std::string &path) const override;
+  void CopyFile(const std::string &src_file_path,
+                const std::string &dst_file_path) const override;
   void CreateDirectory(const std::string &path) const override;
-  void DeleteDirectory(const std::string &path, bool delete_topleveldir) const override;
-  std::string BuildPaxDirectoryPath(RelFileNode rd_node, BackendId rd_backend) const override;
-  std::string BuildPaxFilePath(Relation rel, const std::string &block_id) const override;
+  void DeleteDirectory(const std::string &path,
+                       bool delete_topleveldir) const override;
+  std::string BuildPaxDirectoryPath(RelFileNode rd_node,
+                                    BackendId rd_backend) const override;
+  std::string BuildPaxFilePath(Relation rel,
+                               const std::string &block_id) const override;
 
  private:
   LocalFileSystem() = default;
